requirements.txt

cachelib==0.9.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Caching==2.1.0
Flask-Session==0.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
protobuf==4.21.12
Werkzeug==3.0.1
